+++ app/src/main/java/com/steinwurf/petro/Utils.java
@@ -8,6 +8,7 @@ package com.steinwurf.petro;
 import android.content.ContentUris;
 import android.content.Context;
 import android.database.Cursor;
+import android.graphics.Matrix;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;